[refactor] 차량 선택 페이지 컴포넌트 분리
