+++ frontend/Idle/src/pages/optionPage.jsx
@@ -125,7 +125,6 @@ function OptionPage() {
             navigateTo(car, navigate);
           }}
           detail={"선택이 필요한 페이지로 이동하시겠습니까?"}
-          modalPosition={"modal"}
         />
       ) : null}
       {isDefaultModalOpen ? <DefaultOption setVisible={setIsDefaultModalOpen} /> : <></>}
